updating files structure.
added a few subfolders. moved files to appropriate folders. Still a work in progress.
